+++ test/TestStrlen2.hs
@@ -14,8 +14,8 @@ import           Text.RawString.QQ
 
 testInpPath = "test-data/samples"
 
-testParams = [ ("solver", Just ["z3", "yices", "boolector", "cvc4"])
-             , ("loop-merging", Just ["loopmerge", "loop"])
+testParams = [ ("solver", SpecificValues ["z3", "yices", "boolector", "cvc4"])
+             , ("loop-merging", SpecificValues ["loopmerge", "loop"])
              ]
 
 sugarCube = mkCUBE